chore: improved appkit config types
